# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: WECTI API
  description: >
    API central do sistema de controle de acesso a palestras e eventos.
    O site (aluno e admin) consome exclusivamente esta API -
    nenhum cliente acessa o banco direto. Nao ha app Android dedicado: o
    check-in/check-out é feito via QR code lido pela câmera nativa do
    celular do aluno. Este e um RASCUNHO inicial para alinhar o time
    antes da Fase 1; campos e regras podem mudar conforme o
    desenvolvimento avança.
  version: 0.1.0-draft
servers:
  - url: http://localhost:8080/v1
    description: Ambiente local de desenvolvimento
  - url: https://api.wecti.com.br/v1
    description: Produção (placeholder, ajustar quando definirem o domínio)
...
security:
  - bearerAuth: []
paths:
...
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Lista usuários paginados (somente admin)
      description: >
        Paginado de propósito - a lista de usuários cresce muito quando o
        sistema é liberado pros alunos, devolver tudo de uma vez não
        escala. "busca" filtra por nome, email, RGM ou CPF (contém, sem
        diferenciar maiúsculas/minúsculas no nome/email).
      parameters:
        - name: perfil
          in: query
          schema:
            $ref: '#/components/schemas/Perfil'
        - name: busca
          in: query
          schema:
            type: string
        - name: page
          in: query
          description: Página (começando em 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Página de usuários
          content:
            application/json:
              schema:
                type: object
                properties:
                  conteudo:
                    type: array
                    items:
                      $ref: '#/components/schemas/Usuario'
                  pagina:
                    type: integer
                  tamanho:
                    type: integer
                  total_elementos:
                    type: integer
                  total_paginas:
                    type: integer
    post:
      tags: [Usuarios]
      summary: Cadastra aluno ou convidado avulso
      description: >
        Usado pelo admin direto pelo site quando precisa
        cadastrar alguém na hora, na porta do evento.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovoUsuario'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
  /usuarios/{usuarioId}:
    parameters:
      - name: usuarioId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Usuarios]
      summary: Atualiza um usuário (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovoUsuario'
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '409':
          description: Email ou RGM já usado por outro usuário
    delete:
      tags: [Usuarios]
      summary: Exclui um usuário (somente admin)
      responses:
        '204':
          description: Usuário excluído
        '422':
          description: Usuário possui inscrições e não pode ser excluído
  /usuarios/me:
    get:
      tags: [Usuarios]
      summary: Retorna o usuário autenticado
      responses:
        '200':
          description: Dados do usuário logado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'

    put:
      tags: [Usuarios]
      summary: A própria pessoa corrige o próprio cadastro
      description: >
        Existe porque muito aluno errou nome, RGM ou curso no
        autocadastro e só o admin conseguia arrumar, um a um.


        O id vem do token, nunca do caminho ou do corpo - não há como
        editar o cadastro de outra pessoa. E-mail, perfil, CPF e senha
        NÃO entram aqui: o e-mail é o login (quem digitasse errado ao
        "corrigir" perderia o acesso), o perfil vindo do cliente seria
        escalada de privilégio, e a senha continua pelo fluxo de
        /auth/redefinir-senha. Para admin, rgm e curso são ignorados -
        são campos de aluno.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarPerfil'
      responses:
        '200':
          description: Cadastro atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Nome vazio, ou RGM ausente/fora do formato para aluno
        '409':
          description: RGM já usado por outro usuário

  # ---------------------------------------------------------------------
  # Palestrantes
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Perfil:
      type: string
      # Documentado em maiusculas porque e assim que o Jackson serializa o
      # enum Java por padrao (Usuario.perfil, NovoUsuario.perfil e a claim
      # "perfil" do JWT) - o rascunho original documentava em minusculas,
      # mas nao era isso que a API implementada realmente devolvia/esperava.
      enum: [ADMIN, ALUNO]
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        email:
          type: string
          format: email
        perfil:
          $ref: '#/components/schemas/Perfil'
        rgm:
          type: string
          pattern: '^\d{8}$'
          nullable: true
          description: >
            Registro Geral do Aluno (RGM) - identificador acadêmico da
            faculdade, com 8 dígitos. Obrigatório e único quando
            perfil = aluno; não se aplica a admin.
        cpf:
          type: string
          pattern: '^\d{11}$'
          nullable: true
          description: >
            CPF do admin, 11 dígitos. Obrigatório e único quando
            perfil = admin (é o identificador usado no "esqueci minha
            senha" desse perfil); não se aplica a aluno.
        curso:
          type: string
          nullable: true
          description: >
            Curso do aluno (ex.: "Ciência da Computação") - confirmado
            com o professor (stakeholder do projeto) que é só
            informativo, pra saber quais cursos estão participando das
            palestras. Não afeta pontuação, elegibilidade pra inscrição
            nem nenhuma outra regra de negócio - por isso texto livre,
            opcional, sem unicidade. Não se aplica a admin.
    NovoUsuario:
      type: object
      required: [nome, email, perfil]
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        perfil:
          $ref: '#/components/schemas/Perfil'
        rgm:
          type: string
          pattern: '^\d{8}$'
          nullable: true
          description: >
            Obrigatório quando perfil = aluno. O OpenAPI puro não expressa
            "obrigatório somente se outro campo tiver determinado valor",
            então essa regra precisa ser validada no backend.
        cpf:
          type: string
          pattern: '^\d{11}$'
          nullable: true
          description: >
            Obrigatório quando perfil = admin, pela mesma razão do rgm
            acima - validado no backend.
        curso:
          type: string
          nullable: true
          description: Opcional, só informativo (ver descrição em Usuario.curso).
    AtualizarPerfil:
      type: object
      required: [nome]
      description: >
        Corpo de PUT /usuarios/me - o que a própria pessoa pode corrigir
        sozinha. Bem menor que NovoUsuario (a tela do admin) de propósito:
        e-mail, perfil, CPF e senha ficam de fora. Ver a descrição do
        endpoint para o motivo de cada ausência.
      properties:
        nome:
          type: string
        rgm:
          type: string
          pattern: '^\d{8}$'
          nullable: true
          description: >
            Obrigatório para aluno; ignorado para admin (é campo de
            aluno). Mesma limitação do OpenAPI descrita em
            NovoUsuario.rgm - a regra é validada no backend.
        curso:
          type: string
          nullable: true
          description: Opcional, só informativo (ver descrição em Usuario.curso).
...